+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.44.13] - 2024-10-15
+
 ### Fixed
 
 - Force setProfile to use a valid cost center